back/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
binary==1.0.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.11
django-allauth==0.61.1
django-appconf==1.0.6
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-imagekit==5.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mysqlclient==2.2.6
nest-asyncio==1.6.0
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
phonenumbers==8.13.50
pilkit==3.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.19.2
